% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistandard.R
\name{cistandard}
\alias{cistandard}
\title{Compute the standard confidence interval}
\usage{
cistandard(a, x, y, alpha, sig = NULL)
}
\arguments{
\item{a}{A vector used to specify the parameter of interest}

\item{x}{A known n by p matrix}

\item{y}{A known n-vector of responses}

\item{alpha}{1 - \code{alpha} is the nominal coverage probability of the
confidence interval}

\item{sig}{Standard deviation of the random error.  If a value is not
specified, \code{sig} will be estimated from the data.}
}
\value{
The standard confidence interval
}
\description{
Compute the standard confidence interval
}
\details{
Suppose that \deqn{Y = X \beta + \epsilon} is a random \eqn{n}-vector of
responses, \eqn{X} is a known \eqn{n} by \eqn{p} matrix with linearly
independent columns, \eqn{\beta} is an unknown parameter \eqn{p}-vector and
\eqn{\epsilon} has a multivariate normal distribution with mean vector 0 and
variance \code{sig}^2 times the \eqn{n} by \eqn{n} identity matrix.
Then \code{cistandard} will compute the standard confidence interval for
\code{a}' \eqn{\beta}.

In the example below we use the data set described in Table 7.5
of Box et al. (1963).  A description of the parameter of interest  is given
in Dicsussion 5.8, p.3426 of Kabaila and Giri (2009).
}
\examples{
y <- c(87.2, 88.4, 86.7, 89.2)
x1 <- c(-1, 1, -1, 1)
x2 <- c(-1, -1, 1, 1)
x <- cbind(rep(1, 4), x1, x2, x1*x2)
a <- c(0, 2, 0, -2)
c <- c(0, 0, 0, 1)
t <- 0
res <- cistandard(a, x, y, 0.05, sig = 0.8)
res

}
\references{
Box, G.E.P., Connor, L.R., Cousins, W.R., Davies, O.L., Hinsworth, F.R., Sillitto, G.P. (1963)
The Design and Analysis
of Industrial Experiments, 2nd edition, reprinted. Oliver and Boyd, London.

Kabaila, P. and Giri, K. (2009) Confidence intervals in regression utilizing
prior information.  Journal of Statistical Planning and Inference, 139,
3419 - 3429.
}
